just gonna move the funcs to src in assembly for now so I can combine stuff
